pyproject.toml
[project]
name = "hidwaan"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Jezeniel Zapanta", email = "jezeniel.zapanta@gmail.com" }
]
dependencies = [
    "fastapi>=0.111.0",
    "aiosqlite>=0.20.0",
    "pyjwt>=2.8.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=24.4.2",
    "ipython>=8.25.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/hidwaan"]
